requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-dotenv
pydantic
passlib[bcrypt]
python-jose
faiss-cpu